// RAD-ify/src/types/types.ts
export interface CloudscapeTokenValue {
  light: string;
  dark: string;
}

export interface CloudscapeTheme {
  tokens: Record<string, CloudscapeTokenValue>;
}